Week 6 Class 2 - Final touches
- Added final touches
- Added user check to delete post code to ensure that current user can
ONLY delete his/hers own posts

+++ FeedTableViewController.swift
@@ -17,6 +17,7 @@ class FeedTableViewController: UITableViewController, UIImagePickerControllerDel
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        navigationItem.titleView = UIImageView(image: UIImage(named: "Selfigram-logo"))
         getPosts (block: {})
         
     }
@@ -202,61 +203,64 @@ class FeedTableViewController: UITableViewController, UIImagePickerControllerDel
     // Override to support editing the table view.
     //we've overridden this by placing our camera icon on the top right, otherwise it would show there
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
+        
         if editingStyle == .delete {
             //Delete the row from the data source
             //use the objectID to determine which object the user wants to delete
             
             print("\(indexPath)")
             
-            
-            let post = posts[indexPath.row]
-            let user = PFUser.current()
-            print("Post ID: \(post.objectId)")
-            print("User ID: \(user?.objectId)")
-            
-            //#1 delete from parse in cloud
-            post.deleteInBackground(block: { (success, error) in
+             if let currentUser = PFUser.current() {
                 
-                //if successfull deleting from parse then delete locally
-                if success {
-                    //tableView.deleteRows(at: indexPath, with: UITableViewRowAnimation)
+                let post = posts[indexPath.row]
+                
+                //then we check if the users are the same and if so we can delete
+                if post.user.objectId == currentUser.objectId {
                     
-                    //#2 delete from mirrired array (mirrored to parse)
-                    self.posts.remove(at: indexPath.row)
                     
-                    //#3 now remove from table. it wants an array but we can be explicit
-                    
-                    //indexpath here want row and section
-                    tableView.deleteRows(at: [indexPath], with: .fade)
-                    //refresh is automatic with the above call?
-                    
-                    ////////past post object to isolated method
-                    // but also check to ensure that the user that is deleting the post
-                    //is the owner of the photo, if not do not allow delete HUD message
-                    //PFQuery to find the like activity
-                    if let activityQuery = Activity.query(){
-                        activityQuery.whereKey("post", equalTo: post)
-                        activityQuery.whereKey("user", equalTo: user!)
-                        activityQuery.whereKey("type", equalTo: "like")
-                        activityQuery.findObjectsInBackground(block: { (activities, error) -> Void in
+                    //#1 delete from parse in cloud
+                    post.deleteInBackground(block: { (success, error) in
+                        
+                        //if successfull deleting from parse then delete locally
+                        if success {
+                            //tableView.deleteRows(at: indexPath, with: UITableViewRowAnimation)
                             
+                            //#2 delete from mirrired array (mirrored to parse)
+                            self.posts.remove(at: indexPath.row)
                             
-                            // You should only have one like activity from a user
-                            // but this is code is being safe and checking for one or multiple instances
-                            // and then deleting all of them
-                            if let activities = activities {
-                                for activity in activities {
-                                    activity.deleteInBackground(block: { (success, error) -> Void in
-                                        print("deleted activity")
-                                    })
-                                }
+                            //#3 now remove from table. it wants an array but we can be explicit
+                            
+                            //indexpath here want row and section
+                            tableView.deleteRows(at: [indexPath], with: .fade)
+                            //refresh is automatic with the above call?
+                            
+                            ////////past post object to isolated method
+                            // but also check to ensure that the user that is deleting the post
+                            //is the owner of the photo, if not do not allow delete HUD message
+                            //PFQuery to find the like activity
+                            if let activityQuery = Activity.query(){
+                                activityQuery.whereKey("post", equalTo: post)
+                                activityQuery.whereKey("user", equalTo: currentUser)
+                                activityQuery.whereKey("type", equalTo: "like")
+                                activityQuery.findObjectsInBackground(block: { (activities, error) -> Void in
+                                    
+                                    
+                                    // You should only have one like activity from a user
+                                    // but this is code is being safe and checking for one or multiple instances
+                                    // and then deleting all of them
+                                    if let activities = activities {
+                                        for activity in activities {
+                                            activity.deleteInBackground(block: { (success, error) -> Void in
+                                                print("deleted activity")
+                                            })
+                                        }
+                                    }
+                                })
                             }
-                        })
-                    }
-                    
+                        }
+                    })
                 }
-            })
-            
+            }
         } else if editingStyle == .insert {
             //Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
         }
